Created EventList component.
Moved Map and events to EventList.
Home is now just a welcome page.
Changed logic in routes, navbar and forms to redirect to EventList.
