+++ frontend/src/AllRoutes.tsx
@@ -3,6 +3,7 @@ import { Route, Routes, Navigate } from "react-router-dom";
 import Home from "./Home";
 import UserList from "./UserList";
 import User from "./User";
+import EventList from "./EventList";
 import LoginForm from "./LoginForm";
 import RegistrationForm from "./RegistrationForm";
 import { LoginFormData } from "./interfaces/LoginFormData";
@@ -33,6 +34,8 @@ function AllRoutes({ login, register }: AllRoutesParams) {
       <Route path="/" element={<Home />} />
       <Route path="/users" element={<UserList />} />
       <Route path="/users/:id" element={<User />} />
+      <Route path="/events" element={<EventList />} />
+      {/* <Route path="/events/:id" element={<Event />} /> */}
       <Route path="/login" element={<LoginForm login={login} />} />
       <Route
         path="/register"
